Move LSPS1 client logic into `liquidity/client/lsps1.rs`
